+++ web/js/package.json
@@ -13,5 +13,6 @@
     "express": "^4.19.2",
     "fs": "0.0.1-security",
     "nodemon": "^3.1.4"
-  }
+  },
+  "type": "module"
 }
