+++ web/js/server.js
@@ -1,13 +1,12 @@
-const express = require('express');
-//import express from 'express';
+//const express = require('express');
+import express from 'express';
 const port = 5500;
-
 const app = express();
-const path = require('path');
-//import path from 'path';
-//const __dirname = path.resolve();
 
-const create_story = require('./story_list')
+import fs from 'fs';
+//const path = require('path');
+import path from 'path';
+const __dirname = path.resolve();
 
 app.use(express.static('../public'));
 app.use(express.static('../js'));
@@ -18,10 +17,11 @@ app.get('/', (req, res) => {
 
 app.get('/list', (req, res) => {
     res.sendFile(path.join(__dirname, '../public/html/storyList.html'));
-    const create_story_list = create_story('left',1,2);
-    const contentsContainer = document.createElement('.contents');
-    contentsContainer.innerHTML = create_story_list;
+    //res.sendFile(path.join(__dirname, '../public/html/test.html'));
 });
+app.get('/get_story_list', (req, res) => {
+    res.json(load_novel_data());
+})
 
 app.get('/detail_story/:id', (req, res) => {
     res.sendFile(path.join(__dirname, '../public/html/detailStory.html'));
@@ -31,3 +31,17 @@ app.listen(port, () => {
     console.log(__dirname);
     console.log(`서버가 http://localhost:${port} 에서 실행 중입니다.`);
 });
+
+function load_novel_data() {
+    const datasetFile_path = path.join(__dirname, '../../dataset', 'novel_data.json');
+    console.log(datasetFile_path);
+    try {
+        const data = fs.readFileSync(datasetFile_path, 'utf8');
+        const parsed_json = JSON.parse(data);
+        console.log('정상적으로 파일을 읽었습니다');
+        return parsed_json;
+    } catch (err) {
+        console.error('파일을 읽는 중 오류가 발생했습니다:', err);
+        return null;
+    }
+}
