+++ web/js/test.js
@@ -1,0 +1,5 @@
+import path from 'path';
+const __dirname = path.resolve();
+//const datasetFile_path = path.join(__dirname, '../../dataset', 'novel_data.json');
+const datasetFile_path = '../../dataset/novel_data.json';
+console.log(datasetFile_path);
