+++ web/public/html/detailStory.html
@@ -6,7 +6,6 @@
 	<link rel="stylesheet" href="../css/index.css">
 	<!-- 돌아가기 아이콘-->
 	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-	<script src='../js/index.js'></script>
 </head>
 <body>
 	<a href="/list" style="text-decoration:none;">
@@ -44,5 +43,6 @@
 			</div>
 		</div>
 	</div>
+	<script src='../js/index.js'></script>
 </body>
 </html>
