+++ web/public/html/storyList.html
@@ -1,14 +1,15 @@
 <!doctype html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org" lang="ko">
 <head>
-    <meta charset="utf-8" />
+	<meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>List Page</title>
-    <link rel="stylesheet" href="../css/list.css">
+    <link rel="stylesheet" href="css/list.css">
 </head>
 <body>
 	<a href="/list" style="text-decoration:none;">
 		<div class="header">
-		    <img src="../img/insight-logo.png" alt="로고">
+		    <img src="img/insight-logo.png" alt="로고">
 			<div class="insight-story">
 				<p>Insight Story</p>
 			</div>
@@ -23,7 +24,6 @@
 	</div>
 	<div class="contents">
 	</div>
-	<script src="./story_list.js" type="commonjs">
-	</script>
+	<script type="module" src="story_list.js"></script>
 </body>
 </html>
