+++ web/public/html/test.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+    <head>
+    <title>Story List</title>
+    </head>
+    <body>
+    <script>
+        window.onload = function() {
+        alert("페이지가 생성되었습니다.");
+        };
+    </script>
+    </body>
+</html>
